third_party/WebKit/LayoutTests/inspector/layers/layer-scroll-rects-update.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/layers-test.js"></script>
<script>
function test()
{
    var layers = [
    {
        layerId: "1", offsetX: 0, offsetY: 0, width: 0, height: 0, paintCount: 0
    },
    {
        layerId: "2", offsetX: 0, offsetY: 0, width: 0, height: 0, paintCount: 0, parentLayerId: "1", scrollRects: [
            { rect: {x: 0, y: 10, width: 10, height: 10}, type: "RepaintsOnScroll"}
        ]
    },
    {
        layerId: "3", offsetX: 0, offsetY: 0, width: 50, height: 50, paintCount: 0, parentLayerId: "2", backendNodeId: -1,
        scrollRects: [
            {rect: {x: 10, y: 10, width: 10, height: 10}, type: "WheelEventHandler"},
            {rect: {x: 10, y: 0, width: 10, height: 10}, type: "TouchEventHandler"}
        ]
    },
    {
        layerId: "4", offsetX: 0, offsetY: 0, width: 50, height: 50, paintCount: 0, parentLayerId: "3", backendNodeId: -2, scrollRects: [
            {rect: {x: 0, y: 0, width: 10, height: 10}, type: "TouchEventHandler"}
        ]
    }
    ];
    var changedLayers = [
    {
        layerId: "1", offsetX: 0, offsetY: 0, width: 0, height: 0, paintCount: 0
    },
    {
        layerId: "3", offsetX: 0, offsetY: 0, width: 50, height: 50, paintCount: 0, parentLayerId: "1", backendNodeId: -1, scrollRects: [
            {rect: {x: 0, y: 0, width: 10, height: 10}, type: "WheelEventHandler"},
            {rect: {x: 10, y: 0, width: 10, height: 10}, type: "TouchEventHandler"}
        ]
    },
    {
        layerId: "4", offsetX: 0, offsetY: 0, width: 50, height: 50, paintCount: 0, parentLayerId: "3", backendNodeId: -2, scrollRects: [
            {rect: {x: 0, y: 0, width: 10, height: 10}, type: "TouchEventHandler"},
            {rect: {x: 10, y: 10, width: 10, height: 10}, type: "TouchEventHandler"}
        ]
    }
    ];

    function markScrollRects()
    {
        var root = WebInspector.inspectorView.panel("layers")._layers3DView._rotatingContainerElement;
        Array.prototype.forEach.call(root.querySelectorAll('.scroll-rect'), function(element) {
            element.__unchanged = true;
        });
    }

    function sendLayersAndDump(caption, layers)
    {
        InspectorTest.addResult(caption);
        InspectorTest.layerTreeModel._repopulate(layers);
        InspectorTest.layerTreeModel.dispatchEventToListeners(WebInspector.LayerTreeModel.Events.LayerTreeChanged);
        InspectorTest.dumpViewScrollRects();
        InspectorTest.dumpModelScrollRects();
    }


    WebInspector.inspectorView.showPanel("layers");
    InspectorTest.layerTreeModel._backendNodeIdToNodeId = {
        "-1": 1,
        "-2": 2
    };
    sendLayersAndDump("Initial scroll rectangles", layers);
    markScrollRects();
    sendLayersAndDump("Updated scroll rectangles", changedLayers);
    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
