third_party/WebKit/LayoutTests/inspector/layers/layer-scroll-rects-update-expected.txt

Initial scroll rectangles
View elements dump
scroll-rect: {
    className : "scroll-rect"
    height : "29px"
    left : "0px"
    title : "repaints on scroll"
    top : "29px"
    width : "29px"
}
scroll-rect: {
    className : "scroll-rect"
    height : "29px"
    left : "29px"
    title : "mousewheel event listener"
    top : "29px"
    width : "29px"
}
scroll-rect: {
    className : "scroll-rect"
    height : "29px"
    left : "29px"
    title : "touch event listener"
    top : "0px"
    width : "29px"
}
scroll-rect: {
    className : "scroll-rect"
    height : "29px"
    left : "0px"
    title : "touch event listener"
    top : "0px"
    width : "29px"
}
Model elements dump
{
}
{
    0 : {
        rect : {
            height : 10
            width : 10
            x : 0
            y : 10
        }
        type : "RepaintsOnScroll"
    }
}
{
    0 : {
        rect : {
            height : 10
            width : 10
            x : 10
            y : 10
        }
        type : "WheelEventHandler"
    }
    1 : {
        rect : {
            height : 10
            width : 10
            x : 10
            y : 0
        }
        type : "TouchEventHandler"
    }
}
{
    0 : {
        rect : {
            height : 10
            width : 10
            x : 0
            y : 0
        }
        type : "TouchEventHandler"
    }
}
Updated scroll rectangles
View elements dump
scroll-rect: {
    __unchanged : true
    className : "scroll-rect"
    height : "29px"
    left : "0px"
    title : "mousewheel event listener"
    top : "0px"
    width : "29px"
}
scroll-rect: {
    __unchanged : true
    className : "scroll-rect"
    height : "29px"
    left : "29px"
    title : "touch event listener"
    top : "0px"
    width : "29px"
}
scroll-rect: {
    className : "scroll-rect"
    height : "29px"
    left : "0px"
    title : "touch event listener"
    top : "0px"
    width : "29px"
}
scroll-rect: {
    className : "scroll-rect"
    height : "29px"
    left : "29px"
    title : "touch event listener"
    top : "29px"
    width : "29px"
}
Model elements dump
{
}
{
    0 : {
        rect : {
            height : 10
            width : 10
            x : 0
            y : 0
        }
        type : "WheelEventHandler"
    }
    1 : {
        rect : {
            height : 10
            width : 10
            x : 10
            y : 0
        }
        type : "TouchEventHandler"
    }
}
{
    0 : {
        rect : {
            height : 10
            width : 10
            x : 0
            y : 0
        }
        type : "TouchEventHandler"
    }
    1 : {
        rect : {
            height : 10
            width : 10
            x : 10
            y : 10
        }
        type : "TouchEventHandler"
    }
}
